Noissue - Remove dead code
Remove BuildVariant.h and PluginClient.h. Move their (very few) definitions
into the (just as few) places they were used. Adjust #include directives as
needed. In addition, remove a number of unnecessary #include directives
throughout the plugin.

Remove DLL entry point functions 'OnInstall' and 'OnUpdate'. Remove function
'InitPlugin' that each of them called.

Remove suppression of warning 4996 from PluginStdAfx.h; the compiler does not
emit this warning on our code any longer. Move declaration of '_Module' out
of PluginStdAfx.h and into PluginClass.cpp, which was the only file that
required it.

Move 'UnescapeUrl' out of PluginClientBase.* and into PluginUtil.*.
