+++ src/plugin/PluginClientBase.h
@@ -69,14 +69,4 @@ public:
   static bool PopFirstPluginError(CPluginError& pluginError);
 };
 
-/**
- * Wrapper around Microsoft API 'UrlUnescape'
- *
- * This function has modify-in-place semantics.
- * This behavior matches that of the legacy version of this function declared above.
- * At present, callers of this function have no code to handle error conditions that might arise here.
- * Because there's no error handling, therefore, this masks failures in UrlUnescape.
- */
-void UnescapeUrl(std::wstring& url);
-
 #endif // _PLUGIN_CLIENT_BASE_H_
