+++ src/plugin/PluginWbPassThrough.cpp
@@ -16,16 +16,12 @@
  */
 
 #include "PluginStdAfx.h"
-
 #include "PluginWbPassThrough.h"
-#include "PluginClient.h"
-#include "PluginClientFactory.h"
-#include "PluginFilter.h"
+#include "AdblockPlusClient.h"
 #include "PluginSettings.h"
 #include "PluginClass.h"
-#include "PluginSystem.h"
+#include "PluginUtil.h"
 #include <WinInet.h>
-#include "wtypes.h"
 #include "../shared/Utils.h"
 #include "IeVersion.h"
